+++ package.json
@@ -10,6 +10,7 @@
   "engines": {
     "node": ">=6.0.0"
   },
+  "main": "src/index.js",
   "stripes": {
     "type": "handler",
     "handlerName": "eventHandler",
